# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Voice Router API
  version: 1.0.0
  description: |
    REST бэкенда Voice Router. Базовый адрес http://localhost:8000 (мок: тот же адрес, см. fixtures/README.md).

    WebSocket-каналы (OpenAPI их не описывает, схема событий в ws-events.schema.json):
    - `WS /ws/voice?channel=web&caller_phone=%2B77010000007` — голосовая сессия. Бинарные кадры: вход PCM16 LE 16 kHz mono по 20 мс,
      выход PCM16 LE 24 kHz mono. Текстовые кадры: JSON-события.
    - `WS /ws/supervisor` — поток всех событий всех сессий + события bench/patch/catalog. Только сервер → клиент.

    Телефония: AudioSocket TCP :9092 (AUDIOSOCKET_PORT; протокол в docs/telephony/README.md, диалплан contracts/asterisk/extensions.conf)
    + эндпоинты /api/telephony/*. Сервер шлёт непрерывный поток кадров 0x10 по 20 мс (320 байт), в паузах тишина (keepalive:
    AudioSocket() Asterisk 22 рвёт сокет после 2 с без активности). Конец звонка: 0x00; перевод: AMI Setvar+Redirect, иначе /outcome.
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/route:
    post:
      tags: [meta]
      summary: Разовая маршрутизация текста без сессии (отладка, bench, демо текстом)
      requestBody:
        content:
          application/json:
            schema: {type: object, required: [text], properties: {text: {type: string}, state: {$ref: "dialog-state.schema.json"}}}
      responses:
        "200":
          description: Выход роутера + решение политики
          content:
            application/json:
              schema: {type: object, properties: {router: {$ref: "router-output.schema.json"}, decision: {type: string}, latency_ms: {type: integer}, model: {type: string}}}
  /api/stats:
    get:
      tags: [sessions]
      summary: Агрегаты для панели супервизора
      parameters:
        - {name: since, in: query, schema: {type: string, format: date-time}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/Stats"}}}}
  /api/sessions:
    get:
      tags: [sessions]
      parameters:
        - {name: limit, in: query, schema: {type: integer, default: 50}}
        - {name: channel, in: query, schema: {type: string, enum: [web, phone, text]}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/SessionSummary"}}}}}
  /api/sessions/{session_id}:
    get:
      tags: [sessions]
      parameters: [{name: session_id, in: path, required: true, schema: {type: string}}]
      responses:
        "200":
          description: Сессия с трассировками всех ходов и транскриптом
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SessionSummary"
                  - type: object
                    properties:
                      transcript: {type: array, items: {type: object, properties: {turn: {type: integer}, role: {enum: [client, bot]}, text: {type: string}, lang: {type: string}}}}
                      traces: {type: array, items: {$ref: "trace.schema.json"}}
                      final_state: {$ref: "dialog-state.schema.json"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/sessions/{session_id}/events:
    get:
      tags: [sessions]
      summary: Все события сессии для реплея (JSON Lines, как файлы fixtures/sessions/*.jsonl)
      parameters: [{name: session_id, in: path, required: true, schema: {type: string}}]
      responses:
        "200": {description: OK, content: {application/x-ndjson: {schema: {type: string}}}}
...
  /api/catalog/versions/{version}:
    get:
      tags: [catalog]
      summary: Полный каталог версии (формат data/scenarios.json)
      parameters: [{name: version, in: path, required: true, schema: {type: string}}]
      responses:
        "200": {description: OK, content: {application/json: {schema: {type: object}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/catalog/freeze:
    post:
      tags: [catalog]
      requestBody: {content: {application/json: {schema: {type: object, required: [frozen], properties: {frozen: {type: boolean}}}}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "evolution.schema.json#/$defs/CatalogVersion"}}}}
  /api/catalog/rollback:
    post:
      tags: [catalog]
      requestBody: {content: {application/json: {schema: {type: object, required: [version], properties: {version: {type: string}}}}}}
      responses:
        "200": {description: Новая текущая версия, content: {application/json: {schema: {$ref: "evolution.schema.json#/$defs/CatalogVersion"}}}}
        "409": {$ref: "#/components/responses/Frozen"}
...
components:
  responses:
    NotFound: {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    Frozen: {description: Каталог заморожен, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  schemas:
    Error:
      type: object
      properties: {error: {type: object, properties: {code: {type: string}, message: {type: string}}}}
...
    SessionSummary:
      type: object
      properties:
        session_id: {type: string}
        channel: {type: string, enum: [web, phone, text]}
        caller_phone: {type: [string, "null"]}
        client_id: {type: [string, "null"]}
        started_at: {type: string, format: date-time}
        ended_at: {type: [string, "null"]}
        end_reason: {type: [string, "null"]}
        turns: {type: integer}
        languages: {type: array, items: {type: string}}
        scenarios: {type: array, items: {type: string}}
        handoff_queue: {type: [string, "null"]}
        latency_total_p50: {type: integer}
        catalog_version: {type: string}
        flagged: {type: boolean, description: есть ход с уверенностью ниже порога, SYS_UNCLEAR или ручная пометка}
    Stats:
      type: object
      properties:
        sessions: {type: integer}
        turns: {type: integer}
        by_scenario: {type: array, items: {type: object, properties: {scenario_id: {type: string}, count: {type: integer}}}}
        by_language: {type: object, additionalProperties: {type: integer}}
        by_decision: {type: object, additionalProperties: {type: integer}}
        confidence_histogram: {type: array, items: {type: object, properties: {bucket: {type: string}, count: {type: integer}}}}
        latency_ms: {type: object, description: "p50/p95 по этапам", additionalProperties: {type: object, properties: {p50: {type: integer}, p95: {type: integer}}}}
        speculative_hit_rate: {type: number}
        template_rate: {type: number}
        handoff_rate: {type: number}
        unclear_rate: {type: number}
        flagged_turns: {type: array, items: {type: object, properties: {session_id: {type: string}, turn: {type: integer}, transcript: {type: string}, scenario_id: {type: string}, confidence: {type: number}, decision: {type: string}}}}
...